Issue-06: use double values instead of float
Converted to using double values throughout,
rather than float values. The book's algorithms
expect double precision, and the imprecision of
floats caused problems.

+++ include/camera.h
@@ -12,7 +12,7 @@ namespace raytrace {
 
 class Camera {
 public:
-  Camera(int h_size, int v_size, float fov,
+  Camera(int h_size, int v_size, double fov,
          Matrix<4> transform = identity_matrix())
       : h_size_(h_size), v_size_(v_size), fov_(fov), transform_(transform) {
     compute_pixel_size();
@@ -20,14 +20,14 @@ public:
 
   auto h_size() const -> int { return h_size_; }
   auto v_size() const -> int { return v_size_; }
-  auto fov() const -> float { return fov_; }
+  auto fov() const -> double { return fov_; }
   auto transform() const -> Matrix<4> { return transform_; }
   auto transform(Matrix<4> transform) -> Camera & {
     transform_ = transform;
     return *this;
   }
 
-  auto pixel_size() const -> float { return pixel_size_; }
+  auto pixel_size() const -> double { return pixel_size_; }
 
   auto ray_for_pixel(int x, int y) const -> Ray;
 
@@ -36,11 +36,11 @@ public:
 private:
   int h_size_;
   int v_size_;
-  float fov_;
+  double fov_;
   Matrix<4> transform_;
-  float half_width_;
-  float half_height_;
-  float pixel_size_;
+  double half_width_;
+  double half_height_;
+  double pixel_size_;
 
   void compute_pixel_size();
 };
